Fix catalyst builds, fixes #44
Using WebRTC with catalyst results in this error:
"codesign failed: bundle format is ambiguous (could be app or
framework)"
It turns out this is because of a missing symlink in
"ios-x86_64_arm64-maccatalyst/WebRTC.framework": WebRTC should be
symlinked to point to "Versions/Current/WebRTC".

The binary at "Versions/Current/WebRTC" on the other hand is the wrong
one (no multi arch) and should be removed in favor of
"ios-x86_64_arm64-maccatalyst/WebRTC.framework/WebRTC"

This PR executes this whole shell game.